refactor: rename classes to more meaningful names
