+++ metals-bench/src/main/scala/bench/ServerInitializeBench.scala
@@ -10,7 +10,7 @@ import scala.concurrent.ExecutionContext
 import scala.meta.internal.metals.Buffers
 import scala.meta.internal.metals.logging.MetalsLogger
 import scala.meta.io.AbsolutePath
-import scala.meta.metals.MetalsNewLanguageServer
+import scala.meta.metals.MetalsLanguageServer
 
 import org.eclipse.lsp4j.InitializeParams
 import org.eclipse.lsp4j.InitializedParams
@@ -55,7 +55,7 @@ class ServerInitializeBench {
     val client = new TestingClient(path, buffers)
     MetalsLogger.updateDefaultFormat()
     val ec = ExecutionContext.fromExecutorService(ex)
-    val server = new MetalsNewLanguageServer(ec, sh = sh)
+    val server = new MetalsLanguageServer(ec, sh = sh)
     server.connectToLanguageClient(client)
     val initialize = new InitializeParams
     initialize.setRootUri(path.toURI.toString)
