+++ metals/src/main/scala/scala/meta/metals/lsp/TextDocumentService.scala
@@ -3,18 +3,7 @@ package scala.meta.metals.lsp
 import java.util
 import java.util.concurrent.CompletableFuture
 
-import scala.meta.internal.metals.DidFocusResult
 import scala.meta.internal.metals.HoverExtParams
-import scala.meta.internal.metals.WindowStateDidChangeParams
-import scala.meta.internal.metals.doctor.DoctorVisibilityDidChangeParams
-import scala.meta.internal.metals.findfiles.FindTextInDependencyJarsRequest
-import scala.meta.internal.tvp.MetalsTreeViewChildrenResult
-import scala.meta.internal.tvp.TreeViewChildrenParams
-import scala.meta.internal.tvp.TreeViewNodeCollapseDidChangeParams
-import scala.meta.internal.tvp.TreeViewNodeRevealResult
-import scala.meta.internal.tvp.TreeViewParentParams
-import scala.meta.internal.tvp.TreeViewParentResult
-import scala.meta.internal.tvp.TreeViewVisibilityDidChangeParams
 
 import org.eclipse.lsp4j._
 import org.eclipse.lsp4j.jsonrpc.messages.{Either => JEither}
@@ -23,30 +12,19 @@ import org.eclipse.lsp4j.jsonrpc.services.JsonRequest
 import org.eclipse.{lsp4j => l}
 
 /**
- * Interface which describes current LSP requests and notifications which are
+ * Interface which describes text document LSP requests and notifications which are
  * implemented by Metals.
  *
- * It's equivalent to [[org.eclipse.lsp4j.services.TextDocumentService]] and
- * [[org.eclipse.lsp4j.services.WorkspaceService]] with some additional, metals
- * specific endpoints.
+ * It's equivalent to [[org.eclipse.lsp4j.services.TextDocumentService]] with some return types changed.
  *
  * Each method has a default implementation which throws UnsupportedOperationException. Throwing in this context is ok because:
- * - the default implementation should never be called, because the method is always overridden in [[MetalsLanguageService]]
+ * - the default implementation should never be called, because the method is always overridden in [[scala.meta.internal.metals.MetalsLanguageServer]]
  * - lsp4j wraps each method in try catch and returns an error response to the client
  */
-trait TextDocumentAndWorkspaceService {
+trait TextDocumentService {
+
   @JsonNotification("textDocument/didOpen")
   def didOpen(params: DidOpenTextDocumentParams): CompletableFuture[Unit] =
-    throw new UnsupportedOperationException()
-
-  @JsonNotification("metals/didFocusTextDocument")
-  def didFocus(
-      params: AnyRef
-  ): CompletableFuture[DidFocusResult.Value] =
-    throw new UnsupportedOperationException()
-
-  @JsonNotification("metals/windowStateDidChange")
-  def windowStateDidChange(params: WindowStateDidChangeParams): Unit =
     throw new UnsupportedOperationException()
 
   @JsonNotification("textDocument/didChange")
@@ -61,16 +39,6 @@ trait TextDocumentAndWorkspaceService {
   @JsonNotification("textDocument/didSave")
   def didSave(params: DidSaveTextDocumentParams): CompletableFuture[Unit] =
     throw new UnsupportedOperationException()
-
-  @JsonNotification("workspace/didChangeConfiguration")
-  def didChangeConfiguration(
-      params: DidChangeConfigurationParams
-  ): CompletableFuture[Unit] = throw new UnsupportedOperationException()
-
-  @JsonNotification("workspace/didChangeWatchedFiles")
-  def didChangeWatchedFiles(
-      params: DidChangeWatchedFilesParams
-  ): CompletableFuture[Unit] = throw new UnsupportedOperationException()
 
   @JsonRequest("textDocument/definition")
   def definition(
@@ -200,59 +168,4 @@ trait TextDocumentAndWorkspaceService {
   ): CompletableFuture[util.List[SelectionRange]] =
     throw new UnsupportedOperationException()
 
-  @JsonRequest("workspace/symbol")
-  def workspaceSymbol(
-      params: WorkspaceSymbolParams
-  ): CompletableFuture[util.List[SymbolInformation]] =
-    throw new UnsupportedOperationException()
-
-  @JsonRequest("workspace/executeCommand")
-  def executeCommand(
-      params: ExecuteCommandParams
-  ): CompletableFuture[Object] = throw new UnsupportedOperationException()
-
-  @JsonRequest("workspace/willRenameFiles")
-  def willRenameFiles(
-      params: RenameFilesParams
-  ): CompletableFuture[WorkspaceEdit] =
-    throw new UnsupportedOperationException()
-
-  @JsonNotification("metals/doctorVisibilityDidChange")
-  def doctorVisibilityDidChange(
-      params: DoctorVisibilityDidChangeParams
-  ): CompletableFuture[Unit] = throw new UnsupportedOperationException()
-
-  @JsonRequest("metals/treeViewChildren")
-  def treeViewChildren(
-      params: TreeViewChildrenParams
-  ): CompletableFuture[MetalsTreeViewChildrenResult] =
-    throw new UnsupportedOperationException()
-
-  @JsonRequest("metals/treeViewParent")
-  def treeViewParent(
-      params: TreeViewParentParams
-  ): CompletableFuture[TreeViewParentResult] =
-    throw new UnsupportedOperationException()
-
-  @JsonNotification("metals/treeViewVisibilityDidChange")
-  def treeViewVisibilityDidChange(
-      params: TreeViewVisibilityDidChangeParams
-  ): CompletableFuture[Unit] = throw new UnsupportedOperationException()
-
-  @JsonNotification("metals/treeViewNodeCollapseDidChange")
-  def treeViewNodeCollapseDidChange(
-      params: TreeViewNodeCollapseDidChangeParams
-  ): CompletableFuture[Unit] = throw new UnsupportedOperationException()
-
-  @JsonRequest("metals/treeViewReveal")
-  def treeViewReveal(
-      params: TextDocumentPositionParams
-  ): CompletableFuture[TreeViewNodeRevealResult] =
-    throw new UnsupportedOperationException()
-
-  @JsonRequest("metals/findTextInDependencyJars")
-  def findTextInDependencyJars(
-      params: FindTextInDependencyJarsRequest
-  ): CompletableFuture[util.List[Location]] =
-    throw new UnsupportedOperationException()
 }
